requirements.txt
click>=8.1.7
google-generativeai>=0.3.2
rich>=13.7.0
python-dotenv>=1.0.1
prompt_toolkit>=3.0.43 